[m3] move the current keccak gadget to stacked
This is done in preparation to adding another version of the keccak gadget, that
does not use stacking.

This is not a functional change.
